[issue-55] refactored repository and remote_repository to include from the new module Artifactory::Resource::RepositoryBase
